# public/openapi.yaml
openapi: 3.0.3
info:
  title: ConejoMalo API
  version: '1.0.0'
  description: |
    API pública y de administración para la tienda. Autenticación vía cookie `session`.
    Este documento es la fuente de verdad de los payloads.
servers:
  - url: https://{host}/api
    variables:
      host:
        default: localhost:3000
  - url: http://localhost:3000/api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string, nullable: true }
        fields:
          type: object
          additionalProperties: { type: string }
          nullable: true
    Permission:
      type: object
      properties:
        key: { type: string }
        description: { type: string, nullable: true }
...
    OrderListItem:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [PENDING, PAID, SHIPPED, COMPLETED, CANCELED] }
        customerName: { type: string }
        customerEmail: { type: string }
        itemsCount: { type: integer }
        subtotal: { oneOf: [ { type: string }, { type: number } ] }
        total: { oneOf: [ { type: string }, { type: number } ] }
        createdAt: { type: string, format: date-time }
...
    UserOrderListItem:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [PENDING, PAID, SHIPPED, COMPLETED, CANCELED] }
        itemsCount: { type: integer }
        subtotal: { oneOf: [ { type: string }, { type: number } ] }
        total: { oneOf: [ { type: string }, { type: number } ] }
        createdAt: { type: string, format: date-time }
...
    CategoryAdminLookup:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
...
    AuditLogItem:
      type: object
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        userEmail: { type: string, nullable: true }
        action: { type: string }
        entity: { type: string, nullable: true }
        entityId: { type: string, nullable: true }
        metadata: { type: object, nullable: true }
...
paths:
  /orders/mine:
    get:
      summary: Listar pedidos del usuario autenticado
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, PAID, SHIPPED, COMPLETED, CANCELED] }
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/UserOrderListItem' }
                  page: { type: integer }
                  per: { type: integer }
                  total: { type: integer }
                  totalPages: { type: integer }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
  /admin/categories/search:
    get:
      summary: Búsqueda rápida de categorías (admin)
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CategoryAdminLookup' } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }

  # Admin: Users
...
  /admin/permissions:
    get:
      summary: Listar permisos (admin)
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Permission' } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
  /admin/orders:
    get:
      summary: Listar órdenes (admin)
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, PAID, SHIPPED, COMPLETED, CANCELED] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/OrderListItem' } }
                  page: { type: integer }
                  per: { type: integer }
                  total: { type: integer }
                  totalPages: { type: integer }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
  /admin/audit:
    get:
      summary: Listar auditoría (admin)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/AuditLogItem' } }, nextCursor: { type: string, nullable: true } } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }

  # Admin: Dashboard metrics
...
  /admin/uploads/image:
    post:
      summary: Subir imagen genérica (admin)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                kind: { type: string, enum: [product, category, generic] }
                name: { type: string }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }
        '400': { description: Error de validación }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
security:
  - cookieAuth: []